Change the way recent queries are saved
Recent queries currently save the final string that is sent to the
search engine. This limits what we can do with them later. This changes
this so that we now save the whole filters object together with the
search term so that we can essentially bring the query back to life if
we want to and allow the user to edit it before sending it again.